requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
PyYAML>=6.0
pydantic>=2.7.0
websockets>=12.0

